% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii2nc.R
\name{ascii2nc}
\alias{ascii2nc}
\title{Convert an ASCII file into NetCDF}
\usage{
ascii2nc(
  ascii,
  nc,
  col_value = "value",
  col_date = "date",
  varname,
  col_lon,
  col_lat,
  ...
)
}
\arguments{
\item{ascii}{\code{character}, name of the ascii file.}

\item{nc}{\code{character}, name of the NetCDF file to be created.}

\item{col_value}{\code{character}, name of the value column.}

\item{col_date}{\code{character}, name of the date column. Date values must be
interpretable by \code{\link{as.Date}}.}

\item{varname}{\code{character}, name of the (single) variable in the ASCII file.}

\item{col_lon}{\code{character}, optional, name of the longitude column. If not given
(default) a single site is assumed.}

\item{col_lat}{\code{character}, optional, name of the latitude column. If not given
(default) a single site is assumed.}

\item{...}{Arguments passed to \code{\link[utils]{read.table}} for reading of the ascii
file.}
}
\value{
A NetCDF file is created. This does not follow any conventions! The dimension
names (\code{col_date}, \code{col_lon}, \code{col_lat}) and data are passed as given.
}
\description{
Convert an ASCII file into NetCDF
}
\details{
Data in the ASCII file are expected to be in a \href{https://en.wikipedia.org/wiki/Tidy_data}{tidy format},
i.e. \emph{rows correspond to sample individuals and columns to variables, so that the entry in
the ith row and jth column gives the value of the jth variate as measured or observed on the ith individual}.

It is important to give the correct file specifications as optional arguments passed to
\code{\link[utils]{read.table}} to read the ASCII file (e.g. \code{header = ...} or
\code{sep = ...})! See also examples below.
}
\note{
So far only daily values and a single variable per file are supported
(and expected in the ascii file).

Metadata are kept to a minimum. A coordinate system is not assigned.
Please add them manually if needed.
}
\examples{
# one year of daily data (random values, normally distributed)
data <- data.frame(
  date = seq.Date(as.Date("2000-01-01"), as.Date("2000-12-31"), by = "day"),
  value = rnorm(n = 366, mean = 10, sd = 4)
)

# save as ascii file (here as temporary file)
ascfile <- tempfile()
write.table(data, ascfile, sep = "\t", row.names = FALSE, quote = FALSE)
\dontrun{
  ascii2nc(ascfile, "test.nc", varname = "Precipitation", header = TRUE, sep = "\t")
}

# spatial grid
data <- data.frame(
  date = rep(seq.Date(as.Date("2000-01-01"), as.Date("2000-12-31"), by = "day"), 9),
  value = rnorm(n = 366*9, mean = 10, sd = 4),
  lon = rep(rep(c(30, 30.5, 31), each = 366), 3), lat = rep(c(40,40.5,41), each = 3*366)
)
ascfile <- tempfile()
write.table(data, ascfile, sep = "\t", row.names = FALSE, quote = FALSE)
\dontrun{
  ascii2nc(ascfile, "test_sp.nc", varname = "Precipitation",
         col_lon = "lon", col_lat = "lat", header = TRUE, sep = "\t")
}
}
