\name{asUnitsDataFrame}
\alias{asUnitsDataFrame}
\title{asUnitsDataFrame}
\description{convert data.frame columns to units objects}
\usage{asUnitsDataFrame(data, units = attributes(data)$units)}
\arguments{
  \item{data}{data.frame}
  \item{units}{list with unit string entries
named as a column in \code{data}}
}

\value{data.frame with columns converted to unit objects}

\author{Jannis v. Buttlar, Thomas Wutzler}





