\name{class.ncdf}
\alias{class.ncdf}
\title{transfers R classes to ncdf classes}
\description{Crudely determines the ncdf class from R classes. Only integer, character
and double are implemented yet.}
\usage{class.ncdf(object)}
\arguments{
  \item{object}{object which class should be determined}
}

\value{character string: ncdf class used in the RNetCDF package.}

\author{Jannis v. Buttlar, MPI BGC Jena, Germany, jbuttlar@bgc-jena.mpg.de
Jannis v. Buttlar}



\seealso{\code{\link[RNetCDF]{RNetCDF}}}

