\name{classR2Ncdf}
\alias{classR2Ncdf}
\title{classR2Ncdf}
\description{transfers R classes to ncdf classes
Crudely determines the ncdf class from R classes. Only integer, character
and double are implemented yet.}
\usage{classR2Ncdf(object)}
\arguments{
  \item{object}{object which class should be determined}
}

\value{character string: ncdf class used in the RNetCDF package.}

\author{Jannis v. Buttlar, Thomas Wutzler}





