\name{closeAllNcfiles}
\alias{closeAllNcfiles}
\title{closeAllNcfiles}
\description{close all open ncdf connections
convenience function to close all ncdf connections that are 
currently open. }
\usage{closeAllNcfiles()}

\value{nothing is returned.}

\author{Jannis v. Buttlar, Thomas Wutzler}





