\name{extract.FN.ncdf}
\alias{extract.FN.ncdf}
\title{extract FN ncdf}

\usage{extract.FN.ncdf(path = getwd(), sites, pars, time.ends, dim.borders)}
\arguments{
  \item{path}{
}
  \item{sites}{
}
  \item{pars}{
}
  \item{time.ends}{
}
  \item{dim.borders}{
}
}



\author{Jannis v. Buttlar}





