\name{GapfillNcdfCreateItercube}
\alias{GapfillNcdfCreateItercube}
\title{helper function for GapfillNcdf}

\usage{GapfillNcdfCreateItercube(datacube, iters.n, dims.cycle.length, 
    dims.cycle.id, slices.process, max.cores, slices.n, MSSA, 
    MSSA.blocksize, MSSA.blck.trsh, tresh.fill.dc)}
\arguments{
  \item{datacube}{
}
  \item{iters.n}{
}
  \item{dims.cycle.length}{
}
  \item{dims.cycle.id}{
}
  \item{slices.process}{
}
  \item{max.cores}{
}
  \item{slices.n}{
}
  \item{MSSA}{
}
  \item{MSSA.blocksize}{
}
  \item{MSSA.blck.trsh}{
}
  \item{tresh.fill.dc}{
}
}
\details{helper function for GapfillNcdf that creates the index array used
in the foreach iterations to extract data.     }


\author{Jannis v. Buttlar}



\seealso{\code{\link{GapfillNcdf}}      }

