\name{GapfillNcdfOpenFiles}
\alias{GapfillNcdfOpenFiles}
\title{helper function for GapfillNcdf}

\usage{GapfillNcdfOpenFiles(file.name, var.name, n.steps, print.status)}
\arguments{
  \item{file.name}{
}
  \item{var.name}{
}
  \item{n.steps}{
}
  \item{print.status}{
}
}
\details{helper function for GapfillNcdf that prepares the output files }


\author{Jannis v. Buttlar}



\seealso{\code{\link{GapfillNcdf}}}

