\name{GapfillNcdfSaveResults}
\alias{GapfillNcdfSaveResults}
\title{helper function for GapfillNcdf}
\description{################################## save results #############################}
\usage{GapfillNcdfSaveResults(datacube, reconstruction, args.call.global, 
    slices.without.gaps, dims.cycle.id, ocean.mask, dims.process.id, 
    var.name, dims.process, slices.process, file.name.copy, print.status, 
    process.cells, n.steps, keep.steps)}
\arguments{
  \item{datacube}{
}
  \item{reconstruction}{
}
  \item{args.call.global}{
}
  \item{slices.without.gaps}{
}
  \item{dims.cycle.id}{
}
  \item{ocean.mask}{
}
  \item{dims.process.id}{
}
  \item{var.name}{
}
  \item{dims.process}{
}
  \item{slices.process}{
}
  \item{file.name.copy}{
}
  \item{print.status}{
}
  \item{process.cells}{
}
  \item{n.steps}{
}
  \item{keep.steps}{
}
}
\details{helper function for GapfillNcdf that saves the results ncdf file. }


\author{Jannis v. Buttlar}



\seealso{\code{\link{GapfillNcdf}}  }

