% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNetCDFAtCoords.R
\name{GetNetCDFAtCoords}
\alias{GetNetCDFAtCoords}
\title{Extract timeseries for nearest coordinates from a netcdf file}
\usage{
GetNetCDFAtCoords(
  filename,
  req.coords,
  req.var,
  time.var = "time",
  lon.var = "longitude",
  lat.var = "latitude"
)
}
\arguments{
\item{filename}{Name of the netcdf file}

\item{req.coords}{matrix or dataframe of longitude and latitude coordinates}

\item{req.var}{name of the variable to extract}

\item{time.var}{name of the time variable}

\item{lon.var, lat.var}{names of the longitude and latitude variables in the 
netcdf file}
}
\value{

}
\description{
Extract timeseries for nearest coordinates from a netcdf file
}
\examples{
\dontrun{req.coors = data.frame(
  lon = c(-64.5417, -64.9888, -64.7),
  lat = c(30.0867, 30.6486, 32.47))
GetNetCDFAtCoords("path/to/file.nc",
                  req.coords = req.coords,
                  req.var = "nobs")
}
}
