\name{ncdf.check.file}
\alias{ncdf.check.file}
\title{ncdf check file}
\description{This function checks whether a ncdf file is consistent with the parts of the COARDS/CF
ncdf conventions that the MDI group agreed on}
\usage{ncdf.check.file(file.name, dims = c("longitude", "latitude", 
    "time"), type = "strict", var.check = "single")}
\arguments{
  \item{file.name}{character string: file name to check}
  \item{dims}{
}
  \item{type}{
}
  \item{var.check}{
}
}

\value{logical: (invisible) whether the file passed the check}

\author{Jannis v. Buttlar, MPI BGC Jena, Germany, jbuttlar@bgc-jena.mpg.de
Jannis v. Buttlar}





