\name{ncdf.cut.files}
\alias{ncdf.cut.files}
\title{ncdf cut files}

\usage{ncdf.cut.files(file.names, time.range.out = c(), time.range.file = c(), 
    fun.start = function(x) NULL, fun.end = function(x) NULL, 
    format = "", convert = function(x) chron(paste(x, "15", sep = ""), 
        format = "ymd", out.format = "d-m-y"))}
\arguments{
  \item{file.names}{
}
  \item{time.range.out}{
}
  \item{time.range.file}{
}
  \item{fun.start}{
}
  \item{fun.end}{
}
  \item{format}{
}
  \item{convert}{
}
}



\author{Jannis v. Buttlar}





