\name{NcdfCheckInputSSA}
\alias{NcdfCheckInputSSA}
\title{helper function for Ncdf SSA routines}

\usage{NcdfCheckInputSSA(SSAprocess, ...)}
\arguments{
  \item{SSAprocess}{
}
  \item{\dots}{
}
}
\details{helper function for Ncdf SSA routines that checks the consistency of the 
input paramters.}


\author{Jannis v. Buttlar}



\seealso{\code{\link{GapfillNcdf}}, \code{\link{DecomposeNcdf}} }

