\name{rbindMod}
\alias{rbindMod}
\title{helper function for GapfillNcdf}
\description{###############  combine data from foreach iteration ########################}
\usage{rbindMod(...)}
\arguments{
  \item{\dots}{
}
}
\details{helper function for GapfillNcdf that combines the foreach output
in a convenient way}


\author{Jannis v. Buttlar}



\seealso{\code{\link{GapfillNcdf}}, \code{\link{foreach}}   }

