\name{VisualizeDatacube}
\alias{VisualizeDatacube}
\title{visualize/plot an overview of a ncdf file}

\usage{VisualizeDatacube(data.object, forth.dim = 0, var.name = "auto", 
    parallel = FALSE, n.series = 16, lwd = 2, data = c(), max.cores = 16, 
    ...)}
\arguments{
  \item{data.object}{object to plot: file name or file.con object linking to a ncdf file}
  \item{forth.dim}{position in possible forth dimension (height, spectral band etc) to plot}
  \item{var.name}{character string: name of the variable to plot}
  \item{parallel}{
}
  \item{n.series}{
}
  \item{lwd}{
}
  \item{data}{
}
  \item{max.cores}{
}
  \item{\dots}{
}
}
\details{\if{html}{\out{<img src="../doc/visualize_ncdf_demo.png" alt="image ..visualize_ncdf_demo should be here"/>}}\ifelse{latex}{}{}}


\author{Jannis v. Buttlar}





