\name{VisualizeGapfill}
\alias{VisualizeGapfill}
\title{visualize/plot an overview of a ncdf file}

\usage{VisualizeGapfill(file.orig, file.filled, data.orig = c(), data.filled = c(), 
    n.series = 16, lwd = 2, max.cores = 36, ...)}
\arguments{
  \item{file.orig}{object to plot: file name or file.con object linking to a ncdf file}
  \item{file.filled}{
}
  \item{data.orig}{
}
  \item{data.filled}{
}
  \item{n.series}{
}
  \item{lwd}{
}
  \item{max.cores}{
}
  \item{\dots}{
}
}
\details{\if{html}{\out{<img src="../doc/visualize_ncdf_demo.png" alt="image ..visualize_ncdf_demo should be here"/>}}\ifelse{latex}{}{}}


\author{Jannis v. Buttlar}





