% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNetCDFAtCoords.R
\name{GetNetCDFAtCoords}
\alias{GetNetCDFAtCoords}
\title{Extract nearest neighbour timeseries}
\usage{
GetNetCDFAtCoords(
  filename,
  req.coords,
  req.var,
  time.var = "time",
  lon.var = "longitude",
  lat.var = "latitude",
  verbose = FALSE
)
}
\arguments{
\item{filename}{path to the netcdf file.}

\item{req.coords}{matrix or dataframe of requested longitude (1st column) and
latitude (2nd column) coordinates.}

\item{req.var}{name of the timeseries variable to extract from the netcdf
file.}

\item{time.var}{name of the time variable in the netcdf file.}

\item{lon.var}{name of the longitude variable in the netcdf file.}

\item{lat.var}{name of the latitude variable in the netcdf file.}

\item{verbose}{logical; if \code{TRUE} processing and run time information
are printed to the terminal.}
}
\value{
A data frame of n rows, where n is the product of the number of
  requested coordinates and the time steps in the netcdf file, and five
  columns with the requested longitude, requested latitude, index number of
  the requested coordinate as character vector, time step, and timeseries
  value for the requested variable at each time step.
}
\description{
Extract from a netcdf file the timeseries from the grid cells which are
nearest to a set of requested coordinates. This function is useful if you
want to analyse only a relatively small number of grid cells from a netcdf
file and you do not necessarily need the entire spatial field.
}
\examples{

filename <- system.file("extdata", "test_data.nc", package = "ncdftools")
req.coords = data.frame(lon = c(18, 13), lat = c(59, 52))

GetNetCDFAtCoords(filename, req.coords = req.coords, req.var = "t2m")

## print information on the processing step
\dontrun{
GetNetCDFAtCoords(filename, req.coords = req.coords, req.var = "dummy",
                  verbose = TRUE)
}
}
\author{
Andrew Dolman with contributions from Thomas Muench
}
